seperate article comments into its own component
